Move logic to trait.
