+++ src/AppBundle/Entity/LocalBusiness.php
@@ -19,6 +19,8 @@ use AppBundle\Entity\LocalBusiness\ImageTrait;
 use AppBundle\Enum\FoodEstablishment;
 use AppBundle\Enum\Store;
 use AppBundle\LoopEat\OAuthCredentialsTrait as LoopEatOAuthCredentialsTrait;
+use AppBundle\OpeningHours\OpenCloseInterface;
+use AppBundle\OpeningHours\OpenCloseTrait;
 use AppBundle\Sylius\Product\ProductInterface;
 use AppBundle\Utils\OpeningHoursSpecification;
 use AppBundle\Utils\TimeRange;
@@ -93,7 +95,7 @@ use Vich\UploaderBundle\Mapping\Annotation as Vich;
  * @AssertIsActivableRestaurant(groups="activable")
  * @Enabled
  */
-class LocalBusiness extends BaseLocalBusiness implements CatalogInterface
+class LocalBusiness extends BaseLocalBusiness implements CatalogInterface, OpenCloseInterface
 {
     use Timestampable;
     use SoftDeleteableEntity;
@@ -101,6 +103,7 @@ class LocalBusiness extends BaseLocalBusiness implements CatalogInterface
     use CatalogTrait;
     use FoodEstablishmentTrait;
     use ImageTrait;
+    use OpenCloseTrait;
 
     /**
      * @var int
@@ -217,8 +220,6 @@ class LocalBusiness extends BaseLocalBusiness implements CatalogInterface
 
     protected $preparationTimeRules;
 
-    protected $nextOpeningDateCache = [];
-
     protected $reusablePackagings;
 
     protected $promotions;
@@ -226,8 +227,6 @@ class LocalBusiness extends BaseLocalBusiness implements CatalogInterface
     protected $featured = false;
 
     protected $stripePaymentMethods = [];
-
-    private $timeRanges = [];
 
     /**
      * @Groups({"restaurant"})
@@ -354,168 +353,6 @@ class LocalBusiness extends BaseLocalBusiness implements CatalogInterface
     public function addClosingRule(ClosingRule $closingRule)
     {
         $this->closingRules->add($closingRule);
-    }
-
-    /**
-     * @param \DateTime|null $now
-     * @return boolean
-     */
-    public function hasClosingRuleForNow(\DateTime $now = null)
-    {
-        $closingRules = $this->getClosingRules();
-
-        if (count($closingRules) === 0) {
-            return false;
-        }
-
-        if (!$now) {
-            $now = Carbon::now();
-        }
-
-        // WARNING
-        // This method may be called a *lot* of times (see getAvailabilities)
-        // Thus, we avoid using Criteria, because it would trigger a query every time
-        foreach ($closingRules as $closingRule) {
-            if ($now >= $closingRule->getStartDate() && $now <= $closingRule->getEndDate()) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private function computeTimeRanges(array $openingHours)
-    {
-        if (count($openingHours) === 0) {
-            $this->timeRanges = [];
-            return;
-        }
-
-        foreach ($openingHours as $openingHour) {
-            $this->timeRanges[] = new TimeRange($openingHour);
-        }
-    }
-
-    private function getTimeRanges()
-    {
-        $openingHours = $this->getOpeningHours();
-        if (count($openingHours) !== count($this->timeRanges)) {
-            $this->computeTimeRanges($openingHours);
-        }
-
-        return $this->timeRanges;
-    }
-
-    public function isOpen(\DateTime $now = null)
-    {
-        if (!$now) {
-            $now = Carbon::now();
-        }
-
-        if ($this->hasClosingRuleForNow($now)) {
-
-            return false;
-        }
-
-        foreach ($this->getTimeRanges() as $timeRange) {
-            if ($timeRange->isOpen($now)) {
-
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private function _getNextOpeningDate(\DateTime $now = null)
-    {
-        if (!$now) {
-            $now = Carbon::now();
-        }
-
-        $dates = [];
-
-        foreach ($this->getTimeRanges() as $timeRange) {
-            $dates[] = $timeRange->getNextOpeningDate($now);
-        }
-
-        sort($dates);
-
-        return array_shift($dates);
-    }
-
-    private function _getNextClosingDate(\DateTime $now = null)
-    {
-        if (!$now) {
-            $now = Carbon::now();
-        }
-
-        $dates = [];
-
-        foreach ($this->getTimeRanges() as $timeRange) {
-            $dates[] = $timeRange->getNextClosingDate($now);
-        }
-
-        sort($dates);
-
-        return array_shift($dates);
-    }
-
-    public function getNextOpeningDate(\DateTime $now = null)
-    {
-        if (!$now) {
-            $now = Carbon::now();
-        }
-
-        if (!isset($this->nextOpeningDateCache[$now->getTimestamp()])) {
-
-            $nextOpeningDate = null;
-
-            if ($this->hasClosingRuleForNow($now)) {
-                foreach ($this->getClosingRules() as $closingRule) {
-                    if ($now >= $closingRule->getStartDate() && $now <= $closingRule->getEndDate()) {
-
-                        $nextOpeningDate = $this->_getNextOpeningDate($closingRule->getEndDate());
-                        break;
-                    }
-                }
-            }
-
-            if (null === $nextOpeningDate) {
-                $nextOpeningDate = $this->_getNextOpeningDate($now);
-            }
-
-            $this->nextOpeningDateCache[$now->getTimestamp()] = $nextOpeningDate;
-        }
-
-        return $this->nextOpeningDateCache[$now->getTimestamp()];
-    }
-
-    public function getNextClosingDate(\DateTime $now = null)
-    {
-        if (!$now) {
-            $now = Carbon::now();
-        }
-
-        $nextClosingDates = [];
-        if ($nextClosingDate = $this->_getNextClosingDate($now)) {
-            $nextClosingDates[] = $nextClosingDate;
-        }
-
-        foreach ($this->getClosingRules() as $closingRule) {
-            if ($closingRule->getEndDate() < $now) {
-                continue;
-            }
-            $nextClosingDates[] = $closingRule->getStartDate();
-        }
-
-        $nextClosingDates = array_filter($nextClosingDates, function (\DateTime $date) use ($now) {
-            return $date >= $now;
-        });
-
-        sort($nextClosingDates);
-
-        return array_shift($nextClosingDates);
     }
 
     /**
